Add qbs project files for all examples.
The examples are built and deployed, including the sources and qmake
project files (as in the qmake build).
Note: This effort is about building the examples as part of the Qt
build. They are currently not buildable separately. Making them self-
contained and deploying the qbs project files along with the sources is
to be done later.

Change-Id: Icbcc513394d72a47f59bea7518cbe746f4895520
Reviewed-by: Jake Petroules <jake.petroules@qt.io>